undo moving ITierCheckingBlock
