+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/common/block/blackaurem/BlackAuremBlock.java
@@ -1,6 +1,6 @@
 package com.github.minecraftschurlimods.arsmagicalegacy.common.block.blackaurem;
 
-import com.github.minecraftschurlimods.arsmagicalegacy.api.util.ITierCheckingBlock;
+import com.github.minecraftschurlimods.arsmagicalegacy.common.block.ITierCheckingBlock;
 import com.github.minecraftschurlimods.arsmagicalegacy.common.init.AMBlockEntities;
 import com.github.minecraftschurlimods.arsmagicalegacy.compat.patchouli.PatchouliCompat;
 import net.minecraft.core.BlockPos;
