+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/compat/theoneprobe/EtheriumProbeInfoProvider.java
@@ -1,10 +1,7 @@
 package com.github.minecraftschurlimods.arsmagicalegacy.compat.theoneprobe;
 
 import com.github.minecraftschurlimods.arsmagicalegacy.api.ArsMagicaAPI;
-import com.github.minecraftschurlimods.arsmagicalegacy.api.util.ITierCheckingBlock;
-import com.github.minecraftschurlimods.arsmagicalegacy.common.block.blackaurem.BlackAuremBlock;
-import com.github.minecraftschurlimods.arsmagicalegacy.common.block.celestialprism.CelestialPrismBlock;
-import com.github.minecraftschurlimods.arsmagicalegacy.common.block.obelisk.ObeliskBlock;
+import com.github.minecraftschurlimods.arsmagicalegacy.common.block.ITierCheckingBlock;
 import com.github.minecraftschurlimods.arsmagicalegacy.common.util.TranslationConstants;
 import mcjty.theoneprobe.api.Color;
 import mcjty.theoneprobe.api.IProbeHitData;
@@ -18,7 +15,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
 
 class EtheriumProbeInfoProvider implements IProbeInfoProvider {
     private static final ResourceLocation ID = new ResourceLocation(ArsMagicaAPI.MOD_ID, "etherium");
